Finished coding doubly linked list functions, need to test them

+++ lib/libft/src/linked_list/void_ptr/lstsize.c
@@ -12,9 +12,9 @@
 
 #include "ft_linked_list.h"
 
-int	ft_lstsize(t_list *lst)
+size_t	ft_lstsize(t_list *lst)
 {
-	int	size;
+	size_t	size;
 
 	size = 0;
 	while (lst)
